+++ src/locationTracker/AreaCounters.tsx
@@ -1,11 +1,7 @@
-import { useSelector } from "react-redux";
-import { areaSelector } from "../state/tracker/Selectors";
+import { useSelector } from 'react-redux';
+import { areaSelector } from '../selectors/LogicOutput';
 
-export default function AreaCounters({
-    areaName,
-}: {
-    areaName: string;
-}) {
+export default function AreaCounters({ areaName }: { areaName: string }) {
     const area = useSelector(areaSelector(areaName))!;
     return (
         <span>
